Support updating and refreshing values.
This required some changes to the topic structure, so in order to keep things
clean I split it out from the OpenEnergyMonitor structure, but added a helper
that republishes events in a compatible format to the appropriate topics for
relevant updates.  This enables much greater uses for the Z-Wave devices and
eliminates emon-specific code from the Z-Wave bridge.

Now all updates get published without having their names mapped to the updates
topics.  The README.md and setup.py were updated accordingly.